fix warning for Linux
Former-commit-id: 76b06961f9db3d84f55ca296872f24989e9e293b
Former-commit-id: e70bc377e84fde52d8e4c467b375977e33d4de7a

+++ Frame/SDK/KernelPlugin/AFCClassModule.cpp
@@ -419,7 +419,7 @@ bool AFCClassModule::Load()
 {
     //////////////////////////////////////////////////////////////////////////
     rapidxml::xml_document<> xDoc;
-    int nDataSize = 0;
+    size_t nDataSize = 0;
 
     std::string strFile = pPluginManager->GetConfigPath() + msConfigFileName;
